Example
I was messing around at work and decided to put together something similar to our project. I used a table from work so it is not exactly what we are doing, but it might help with the front end.  I was able to show the database table in a grid and was able to add the edit, delete, and add buttons.  I was also able to pass the table id's down to javascript functions.  I did not have time to figure out how to pop a pop up window for the edit and add functions.  It might be useful if you run into any problems with the front end.